requirements.txt
flask==3.0.0
flask-cors==4.0.0
Pillow==10.0.0
PyMuPDF==1.23.8
gunicorn==21.2.0
